Create authorization for 'flashcards'

+++ EnglishJourney.Application/Flashcard/Commands/DeleteCategory/DeleteCategoryCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using EnglishJourney.Domain.Entities;
+﻿using EnglishJourney.Domain.Constants;
+using EnglishJourney.Domain.Entities;
 using EnglishJourney.Domain.Exceptions;
 using EnglishJourney.Domain.Interfaces;
 using MediatR;
@@ -6,7 +7,8 @@ using Microsoft.Extensions.Logging;
 
 namespace EnglishJourney.Application.Flashcard.Commands.DeleteCategory
 {
-    public class DeleteCategoryCommandHandler(IFlashcardRepository repository, ILogger<DeleteCategoryCommandHandler> logger)
+    public class DeleteCategoryCommandHandler(IFlashcardRepository repository, ILogger<DeleteCategoryCommandHandler> logger,
+        IEnglishJourneyAuthorizationService authorizationService)
         : IRequestHandler<DeleteCategoryCommand>
     {
         public async Task Handle(DeleteCategoryCommand request, CancellationToken cancellationToken)
@@ -15,6 +17,9 @@ namespace EnglishJourney.Application.Flashcard.Commands.DeleteCategory
 
             var category = await repository.GetFlashardCategoryById(request.Id);
             if (category == null) throw new NotFoundException(nameof(FlashcardCategory), request.Id.ToString());
+
+            if (!authorizationService.AuthorizeFlashcard(category, ResourceOperation.Delete))
+                throw new ForbidException();
 
             await repository.DeleteFlashcardCategory(category);
         }
